feat:added-new-status-to-purchaseRequests

+++ app/Http/Controllers/ChartOfAccountsController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Imports\ChartOfAccountsImport;
+use App\Exports\ChartOfAccountsExport;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreChartOfAccountsRequest;
 use App\Http\Requests\PutChartOfAccountsRequest;
+use App\Models\ChartOfAccounts;
 use App\Services\ChartOfAccountsService as ChartOfAccountsService;
 
 class ChartOfAccountsController extends Controller
@@ -41,6 +43,12 @@ class ChartOfAccountsController extends Controller
 
     public function destroy($id)
     {
+        if(ChartOfAccounts::where('parent', $id)->exists())
+        {
+            return response()->json([
+                'erro' => 'Este plano de contas está associado a outros planos de contas é necessário apagar e/ou alterar suas dependências antes de apagá lo.'
+            ], 422);
+        }
         $chartOfAccounts = $this->chartOfAccountsService->deleteChartOfAccounts($id);
         return response('');
     }
@@ -50,4 +58,23 @@ class ChartOfAccountsController extends Controller
         (new ChartOfAccountsImport)->import(request()->file('import_file'));
         return response('');
     }
+
+    public function export(Request $request)
+    {
+        if(array_key_exists('exportFormat', $request->all()))
+        {
+            if($request->all()['exportFormat'] == 'csv')
+            {
+                return (new ChartOfAccountsExport($request->all()))->download('planoDeContas.csv', \Maatwebsite\Excel\Excel::CSV, ['Content-Type' => 'text/csv']);
+            }
+        }
+        return (new ChartOfAccountsExport($request->all()))->download('planoDeContas.xlsx', \Maatwebsite\Excel\Excel::XLSX);
+    }
+
+
+    public function allChartOfAccounts(Request $request)
+    {
+        return $this->chartOfAccountsService->allChartOfAccounts($request->all());
+    }
+
 }
