+++ app/Http/Controllers/CostCenterController.php
@@ -3,9 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Imports\CostCentersImport;
+use App\Exports\CostCentersExport;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreCostCenterRequest;
 use App\Http\Requests\PutCostCenterRequest;
+use App\Models\CostCenter;
+use App\Models\PaymentRequest;
 use App\Services\CostCenterService as CostCenterService;
 
 class CostCenterController extends Controller
@@ -22,6 +25,12 @@ class CostCenterController extends Controller
     {
         return $this->costCenterService->getAllCostCenter($request->all());
     }
+
+    public function costCenterFilterUser(Request $request)
+    {
+        return $this->costCenterService->costCenterFilterUser($request->all());
+    }
+
 
     public function show($id)
     {
@@ -41,6 +50,20 @@ class CostCenterController extends Controller
 
     public function destroy($id)
     {
+        if(CostCenter::where('parent', $id)->exists())
+        {
+            return response()->json([
+                'erro' => 'Este centro de custo está associado a outros centro de custos é necessário apagar e/ou alterar suas dependências antes de apagá lo.'
+            ], 422);
+        }
+
+        if(PaymentRequest::where('cost_center_id', $id)->exists())
+        {
+            return response()->json([
+                'erro' => 'Este centro de custo está associado a uma ou várias solicitações de pagamento.'
+            ], 422);
+        }
+
         $costCenter = $this->costCenterService->deleteCostCenter($id);
         return response('');
     }
@@ -49,5 +72,23 @@ class CostCenterController extends Controller
     {
         (new CostCentersImport)->import(request()->file('import_file'));
         return response('');
+    }
+
+
+    public function export(Request $request)
+    {
+        if(array_key_exists('exportFormat', $request->all()))
+        {
+            if($request->all()['exportFormat'] == 'csv')
+            {
+                return (new CostCentersExport($request->all()))->download('centrosDeCusto.csv', \Maatwebsite\Excel\Excel::CSV, ['Content-Type' => 'text/csv']);
+            }
+        }
+        return (new CostCentersExport($request->all()))->download('centrosDeCusto.xlsx', \Maatwebsite\Excel\Excel::XLSX);
+    }
+
+    public function allCostCenters(Request $request)
+    {
+        return $this->costCenterService->allCostCenters($request->all());
     }
 }
